backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
SQLAlchemy>=2.0.0
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
pydantic[email]>=2.7.0
pydantic-settings>=2.3.0
alembic>=1.13.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==3.2.2
python-multipart>=0.0.9
openai>=1.30.0
numpy-financial>=1.0.0
